backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langgraph==0.0.8
langchain==0.0.350
langchain-openai==0.0.2
pymupdf==1.23.8
python-docx==1.1.0
pandas
python-multipart==0.0.6
aiofiles==23.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
